Move d2 parser to folder
