+++ d2_sync_report/data/repositories/d2_logs_parser/job_reducer_types.py
